refactor(python): split workflow.rs into venv + base_env
285-line workflow.rs split: venv (init/freeze/sync of .venv) and
base_env (base pip workflow). Public API preserved.
